moves Character class to class file
